+++ src/Runner/Containers/LabelsBuilder.php
@@ -5,6 +5,7 @@
 namespace Ktomk\Pipelines\Runner\Containers;
 
 use Ktomk\Pipelines\Runner\Runner;
+use Ktomk\Pipelines\Utility\App as UtilityApp;
 use Ktomk\Pipelines\Value\Prefix;
 
 /**
@@ -77,7 +78,7 @@ class LabelsBuilder
     {
         $labels = array();
 
-        $labels['pipelines.prefix'] = $this->prefix;
+        $labels['pipelines.prefix'] = UtilityApp::UTILITY_NAME === $this->prefix ? null : $this->prefix;
         $labels['pipelines.role'] = $this->role;
         $labels['pipelines.project.name'] = $this->project;
         $labels['pipelines.project.path'] = $this->projectDirectory;