+++ src/Runner/StepRunner.php
@@ -291,6 +291,7 @@ class StepRunner
             array(
                 $network,
                 '-i', '--name', $container->getName(),
+                $container->obtainLabelOptions(),
                 $env->getArgs('-e'),
                 ArgsBuilder::optMap('-e', $step->getEnv(), true),
                 $mountWorkingDirectory, '-e', 'BITBUCKET_CLONE_DIR=/app',