+++ CHANGELOG.md
@@ -6,6 +6,8 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [unreleased]
+### Fixed
+- Regression of missing labels for step containers since 0.0.43
 
 ## [0.0.45] - 2020-07-09
 ### Added